completed

+++ app/src/main/java/com/gourav/currencyconverter/data/models/Rates.kt
@@ -1,3 +1,5 @@
 package com.gourav.currencyconverter.data.models
 
-data class Rates(val currencyName: String, val amount: Double)
+import java.math.BigDecimal
+
+data class Rates(val currencyName: String, val amount: BigDecimal)
